@@ -13,21 +13,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.0.4",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "2.0.0-beta.12",
-    "@angular/forms": "^5.2.0",
-    "@angular/material": "^5.0.4",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/platform-server": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@auth0/angular-jwt": "^1.0.0",
-    "@ngx-translate/core": "^9.1.1",
-    "@ngx-translate/http-loader": "^2.0.1",
+    "@angular/animations": "7.2.1",
+    "@angular/cdk": "^7.2.1",
+    "@angular/common": "7.2.1",
+    "@angular/compiler": "7.2.1",
+    "@angular/core": "7.2.1",
+    "@angular/flex-layout": "7.0.0-beta.23",
+    "@angular/forms": "7.2.1",
+    "@angular/material": "^7.2.1",
+    "@angular/platform-browser": "7.2.1",
+    "@angular/platform-browser-dynamic": "7.2.1",
+    "@angular/router": "7.2.1",
+    "@auth0/angular-jwt": "^2.1.0",
+    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/http-loader": "^4.0.0",
+    "@types/zxcvbn": "^4.4.0",
     "ajv": "^6.0.1",
     "angular-formio": "^3.9.5",
     "angular-password-strength-meter": "0.0.4",
@@ -36,26 +36,26 @@
     "core-js": "^2.5.3",
     "hammerjs": "2.x",
     "jasmine": "^2.99.0",
-    "jquery": "^1.11.3",
+    "jquery": "^3.3.1",
     "lodash": "4.x",
     "ng2-tooltip-directive": "^2.1.6",
     "ngx-pagination": "^3.2.1",
     "ngx-webstorage-service": "^3.1.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.3.3",
     "ts-helpers": "^1.1.2",
-    "zxcvbn": "^4.4.2"
+    "tslib": "^1.9.0"
   },
   "devDependencies": {
-    "npm": "^6.4.1",
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular/cli": "7.2.2",
+    "@angular/compiler-cli": "7.2.1",
     "@babel/core": "^7.1.0",
     "@babel/preset-env": "^7.1.0",
     "@types/jasmine": "^2.5.53",
     "@types/node": "^9.3.0",
-    "angular-ide": "^0.9.39",
+    "angular-ide": "^0.9.55",
     "babel-preset-env": "^1.7.0",
-    "codelyzer": "~4.0.2",
+    "codelyzer": "~4.5.0",
     "intl": "^1.2.5",
     "jasmine-core": "^2.99.1",
     "jasmine-spec-reporter": "4.2.1",
@@ -67,12 +67,12 @@
     "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-remap-istanbul": "^0.6.0",
-    "karma-webpack": "^3.0.5",
-    "protractor": "~5.2.2",
+    "npm": "^6.4.1",
+    "protractor": "^5.4.2",
     "puppeteer": "^1.8.0",
     "ts-node": "4.1.0",
-    "tslint": "^5.9.1",
-    "typescript": "~2.4.2",
-    "zone.js": "^0.8.20"
+    "tslint": "^5.12.1",
+    "typescript": "3.2.4",
+    "zone.js": "^0.8.28"
   }
 }
