@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {GenericDataService} from './genericdata.service';
 import {HttpClient} from '@angular/common/http';
 import {AppConfigService} from './appconfig.service';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Configuration} from '../model/configuration';
 
 @Injectable()
