@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {GenericDataService} from "./genericdata.service";
 import {AppConfigService} from "./appconfig.service";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs";
 import {User} from "../model";
 import {HttpClient} from '@angular/common/http';
 
