@@ -7,7 +7,7 @@ import {UserService} from '../../../service/user.service';
 import {BaseComponent} from '../../common/basecomponent/base.component';
 import {Component, Input, OnInit} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {Observable} from "rxjs/Observable";
+import {Observable, of} from "rxjs";
 import {CacheService} from "../../../service/cache.service";
 import {UserDataService} from "../../../service/userdata.service";
 import {isNullOrUndefined} from "util";
@@ -99,7 +99,7 @@ export class UserPrivilegesComponent extends BaseComponent implements OnInit {
 
   public getDomainName(domainId: number): Observable<string> {
         if (this.domainCache.hasData(domainId)) {
-            return Observable.of(this.domainCache.getData(domainId).name);
+            return of(this.domainCache.getData(domainId).name);
         } else {
             return this.domainService.getOne(domainId).map((domain) => {this.domainCache.setData(domainId, domain); return domain.name})
                 .shareReplay(1).take(1);
