@@ -7,7 +7,7 @@ import {AppConfigService, ConfigurationService} from "./service";
 import {HttpClient, HttpHandler} from "@angular/common/http";
 import {TranslateService, TranslateModule, TranslateLoader} from "@ngx-translate/core";
 import {TranslateFakeLoader} from "@ngx-translate/core";
-import {Observable} from "rxjs";
+import {Observable, of} from "rxjs";
 import {Configuration} from "./model/configuration";
 
 class MockConfigurationService{
@@ -22,11 +22,11 @@ class MockConfigurationService{
     }
 
     public getConfiguration():Observable<Configuration>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 
     public updateConfiguration(configuration:Configuration):Observable<any>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 }
 
