@@ -3,7 +3,7 @@ import {inject, TestBed} from '@angular/core/testing';
 import {ConfigurationService} from './configuration.service';
 import {AppConfigService} from "./appconfig.service";
 import {HttpClient, HttpHandler} from "@angular/common/http";
-import {Observable} from "rxjs";
+import {Observable, of} from "rxjs";
 import {Configuration} from "../model/configuration";
 
 class MockConfigurationService{
@@ -18,11 +18,11 @@ class MockConfigurationService{
     }
 
     public getConfiguration():Observable<Configuration>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 
     public updateConfiguration(configuration:Configuration):Observable<any>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 }
 
