@@ -4,7 +4,7 @@ import { ContentDisplayService } from './content-display.service';
 import {ConfigurationService} from "./configuration.service";
 import {HttpClient, HttpHandler} from "@angular/common/http";
 import {AppConfigService} from "./appconfig.service";
-import {Observable} from "rxjs";
+import {Observable, of} from "rxjs";
 import {Configuration} from "../model/configuration";
 
 class MockConfigurationService{
@@ -19,11 +19,11 @@ class MockConfigurationService{
     }
 
     public getConfiguration():Observable<Configuration>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 
     public updateConfiguration(configuration:Configuration):Observable<any>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 }
 
