@@ -2,7 +2,7 @@
 
 import { TestBed, async, inject } from '@angular/core/testing';
 import { TagService } from './tag.service';
-import {Observable} from "rxjs";
+import {Observable, of} from "rxjs";
 import {Configuration} from "../model/configuration";
 import {HttpClient, HttpHandler} from "@angular/common/http";
 import {AppConfigService} from "./appconfig.service";
@@ -19,11 +19,11 @@ class MockConfigurationService{
     }
 
     public getConfiguration():Observable<Configuration>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 
     public updateConfiguration(configuration:Configuration):Observable<any>{
-        return Observable.of<Configuration>();
+        return of<Configuration>();
     }
 }
 
