@@ -2,7 +2,7 @@ import {async, ComponentFixture, getTestBed, TestBed} from '@angular/core/testin
 
 import { RatingExtendedComponent } from './rating-extended.component';
 import {AppsService} from "../../service";
-import {Observable} from "rxjs";
+import {Observable, of} from "rxjs";
 import {Rate} from "../../model";
 import {HttpResponse} from "selenium-webdriver/http";
 import {TooltipModule} from "ng2-tooltip-directive";
@@ -10,11 +10,11 @@ import {Pipe, PipeTransform} from "@angular/core";
 
 class MockAppService{
   public getAppRateByUrl(urlPath: string): Observable<Rate> {
-    return Observable.of(new Rate(1, 1, new Map()));
+    return of(new Rate(1, 1, new Map()));
   }
 
   public setMyAppRateByUrl(urlPath: string): Observable<any> {
-    return Observable.of(HttpResponse.prototype);
+    return of(HttpResponse.prototype);
   }
 }
 
