@@ -1,5 +1,5 @@
 import { Directive, OnInit, EventEmitter, Output, OnDestroy, Input } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { SortService } from '../../service/sort.service';
 
