@@ -15,7 +15,7 @@ import 'core-js/es6/map';
 import 'core-js/es6/set';
 
 import 'core-js/es6/reflect';
-import 'core-js/es7/reflect';
+
 
 import 'intl';
 import 'zone.js/dist/zone';
